Remove blackbox probe code
